+++ lib/base/connection.js
@@ -23,11 +23,11 @@ const Readable = require('stream').Readable;
 const Queue = require('denque');
 const SqlString = require('sqlstring');
 const { createLRU } = require('lru.min');
-const PacketParser = require('./packet_parser.js');
-const Packets = require('./packets/index.js');
-const Commands = require('./commands/index.js');
-const ConnectionConfig = require('./connection_config.js');
-const CharsetToEncoding = require('./constants/charset_encodings.js');
+const PacketParser = require('../packet_parser.js');
+const Packets = require('../packets/index.js');
+const Commands = require('../commands/index.js');
+const ConnectionConfig = require('../connection_config.js');
+const CharsetToEncoding = require('../constants/charset_encodings.js');
 
 let _connectionId = 0;
