+++ lib/base/pool.js
@@ -3,9 +3,9 @@
 const process = require('process');
 const SqlString = require('sqlstring');
 const EventEmitter = require('events').EventEmitter;
-const PoolConnection = require('./pool_connection.js');
+const PoolConnection = require('../pool_connection.js');
 const Queue = require('denque');
-const BaseConnection = require('./base_connection.js');
+const BaseConnection = require('./connection.js');
 
 function spliceConnection(queue, connection) {
   const len = queue.length;
