+++ lib/base/pool_connection.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const BaseConnection = require('./base_connection.js');
+const BaseConnection = require('./connection.js');
 
 class BasePoolConnection extends BaseConnection {
   constructor(pool, options) {
