+++ lib/connection.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const BaseConnection = require('./base_connection.js');
+const BaseConnection = require('./base/connection.js');
 
 class Connection extends BaseConnection {
   promise(promiseImpl) {
-    const PromiseConnection = require('./promise_connection.js');
+    const PromiseConnection = require('./promise/connection.js');
     return new PromiseConnection(this, promiseImpl);
   }
 }
