+++ lib/pool.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const BasePool = require('./base_pool.js');
+const BasePool = require('./base/pool.js');
 
 class Pool extends BasePool {
   promise(promiseImpl) {
-    const PromisePool = require('./promise_pool.js');
+    const PromisePool = require('./promise/pool.js');
     return new PromisePool(this, promiseImpl);
   }
 }
