+++ lib/pool_connection.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const BasePoolConnection = require('./base_pool_connection.js');
+const BasePoolConnection = require('./base/pool_connection.js');
 
 class PoolConnection extends BasePoolConnection {
   promise(promiseImpl) {
-    const PromisePoolConnection = require('./promise_pool_connection.js');
+    const PromisePoolConnection = require('./promise/pool_connection.js');
     return new PromisePoolConnection(this, promiseImpl);
   }
 }
