+++ lib/promise/connection.js
@@ -1,10 +1,10 @@
 'use strict';
 
 const EventEmitter = require('events').EventEmitter;
-const PromisePreparedStatementInfo = require('./promise_prepared_statement_info.js');
+const PromisePreparedStatementInfo = require('./prepared_statement_info.js');
 const makeDoneCb = require('./make_done_cb.js');
 const inheritEvents = require('./inherit_events.js');
-const BaseConnection = require('./base_connection.js');
+const BaseConnection = require('../base/connection.js');
 
 class PromiseConnection extends EventEmitter {
   constructor(connection, promiseImpl) {
