+++ lib/promise/inherit_events.js
