+++ lib/promise/make_done_cb.js
