+++ lib/promise/pool.js
@@ -2,9 +2,9 @@
 
 const EventEmitter = require('events').EventEmitter;
 const makeDoneCb = require('./make_done_cb.js');
-const PromisePoolConnection = require('./promise_pool_connection.js');
+const PromisePoolConnection = require('./pool_connection.js');
 const inheritEvents = require('./inherit_events.js');
-const BasePool = require('./base_pool.js');
+const BasePool = require('../base/pool.js');
 
 class PromisePool extends EventEmitter {
   constructor(pool, thePromise) {
