+++ lib/promise/pool_connection.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const PromiseConnection = require('./promise_connection.js');
-const BasePoolConnection = require('./base_pool_connection.js');
+const PromiseConnection = require('./connection.js');
+const BasePoolConnection = require('../base/pool_connection.js');
 
 class PromisePoolConnection extends PromiseConnection {
   constructor(connection, promiseImpl) {
