+++ lib/promise/prepared_statement_info.js
